Enable build under NixOS
Enable the build under NixOS by explicitly calling the perl interpreter when building the SWH LADSPA plugins. The reason is that NixOS does not have much of a `/usr/bin` directory:

```
$ which perl
/run/current-system/sw/bin/perl
```

Hence the build failed under NixOS when relying on the first hash-bang-slash line of the `makestub.pl` script which wants to be interpreted by `/usr/bin/perl -w`.