Add status information for worker threads By allowing the ability to query for the status of worker threads, it grants us stronger and more reliable synchronization by telling us what the worker thread is doing at any given moment. This also fixes the issue where resetting the job queue would race with audio threads currently inside the JobQueue::run call.
The inlining of the master audio thread
processing jobs in startAndWaitForJobs was removed as I didn't
want to duplicate my updates to the status in multiple
functions. The inlining also made the synchronization
between the master audio thread and it's worker threads
harder to understand.

I also have strong reason to doubt that
inlining the master audio thread helps with latency in any
significant way. One thread being inlined will not make the
grand difference in latency, and most of the latency issues
are very likely just a symptom of poor design.