Activate volume envelope by default (#6121)
The deactivated volume envelope leads to clicky sounds. This is fixed by activating it by default for instruments that are added to the project as well as in the default project.

The following changes are made to the envelopes:
* The amount of the volume envelope is set to 1, i.e. it is activated by default.
* The amounts of the cutoff and resonance envelopes are kept at 0 so that users have to opt-in to them. This might also potentially save resources.
* The amount for sustain is changed from 0.5 to 1. This way users start with a full sustained sound.
* The amount for hold is set to 0.

The reasoning for the changes is to enable users to quickly dial in some well known envelopes, especially if they only know ADSR envelopes and don't known how an AHDSR envelope works. A user that wants to dial in some pluck envelope might try to turn the sustain down to 0. However, if the "hold" parameter was set to 0.5 by default it would get in the way and users had to learn that they have to turn the "hold" down to 0. With the new set of default values they can discover the "hold" feature without it getting into the way with simpler approaches.

## Technical details
The constructor of `EnvelopeAndLfoParameters` was extended with the new parameter `defaultAmount` so that we are able to quickly set amounts of 1 or 0 for the volume, cutoff and resonance envelopes. See the changes in `InstrumentSoundShaping`.

Adjust `default.mpt` according to the code changes so that not only newly added instruments have the better defaults.