Initialize volume envelopes by drop target
Implement the following behavior:
* If an instrument is dropped into the song editor it is initialized with the new default values of the volume envelope that ensure that notes do not click.
* If an instrument is added from the instrument list using the context menu then it is added to the song editor and the same behavior applies.
* If an instrument is dropped into the pattern editor it is initialized with the "legacy" volume envelope. This envelope will lead to clicks but enables samples, kicks, etc. to ring out like they did before.

Technical details
------------------
Add the method `TrackContainerView::handleDropEvent` which takes a `QDropEvent` and a `bool` that indicates whether the added instrument must be reset to the legacy volume envelope. `PatternEditor::dropEvent` delegates to this new method with the boolean set to `true` because we want the legacy behavior if an instrument is dropped onto the pattern editor. `TrackContainerView::dropEvent` delegates with the boolean set to `false` because we want the new and improved volume envelopes when used in the song editor context.

`TrackContainerView::handleDropEvent` contains the previous code of `TrackContainerView::dropEvent` plus the extension that implements the potential resetting to the legacy volume envelopes.

Add the method `resetToLegacyVolumeEnvelope` to the following classes:
* `Instrument`: delegates to `InstrumentTrack`
* `InstrumentTrack`: delegates to `InstrumentSoundShaping`
* `InstrumentSoundShaping`: contains the actual implementation using the extended interface of `EnvelopeAndLfoParameters`.

Add two methods to `EnvelopeAndLfoParameters`:
* Setter for predelay, attack, hold, decay, sustain and release
* Setter for the envelope amount

Extend `PluginDescWidget::openInNewInstrumentTrack` so that it adds the instrument using the new default envelopes.

Extend `InstrumentLoaderThread` so that it knows if the instrument must be reset to the legacy volume envelope.