Conditional sound shaping tab
Only add the sound shaping tab for instruments that are not single streamed. This removes the "Nothing to see here" message and makes the GUI more concise. Examples for single streamed instruments are VeSTige, OpulenZ and LB302.

Technical details
------------------
Remove all code related to showing the message "Envelopes, LFOs and filters are not supported by the current instrument." from `InstrumentSoundShapingView`.

Initialize `m_ssView` in `InstrumentTrackWindow` to `nullptr`. Only initialize it for instruments that are not single streamed. Add `nullptr` checks to all references. Use a variable to increment the index for calls to `addTab` because the addition of `m_ssView` now becomes conditional.