Add single item every millisecond
Given that buildSearchTree is actually very cheap (~600 instructions last I
profiled it), the CPU is most likely fast enough to run this multiple times a millisecond.
However, we can't call this function as fast as the CPU is going as that
would boggle down the event loop and freeze the program. If we were to
call buildSearchTree once every millisecond, this should be
fine for most CPUs. This also helps in terms of getting results from the
search relatively quickly.