Simplification and optimization
Simplify the statement in the log call as proposed in a code review.

Furthermore define the minimum float as a const variable in the method. This is done because the "Compiler Explorer" at https://godbolt.org/ implies that std::numeric_limits<float>::min() is an actual function call when used directly. When defined as a const variable it is compiled as a const symbol in the code.